financial-agent-platform/requirements.txt
langgraph
langchain-openai
langchain-core
python-dotenv
ollama
langchain-ollama
langchain_community
langchain-chroma
langchain-huggingface
huggingface-hub
pypdf
psycopg2-binary
sqlalchemy 
pgvector 
langchain-postgres 
langchain-text-splitters 
sentence-transformers
pandas
requests
sec_edgar_downloader
bs4
lxml
docker